+++ vite-project/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.23.1",
     "styled-components": "^6.1.11"
   },
   "devDependencies": {
