+++ vite-project/src/components/popUps/popUser/popUser.jsx
@@ -1,3 +1,5 @@
+import { Link } from "react-router-dom";
+
 const PopUser = () => {
   return (
     <div className="pop-exit" id="popExit">
@@ -9,10 +11,10 @@ const PopUser = () => {
           <form className="pop-exit__form" id="formExit" action="#">
             <div className="pop-exit__form-group">
               <button className="pop-exit__exit-yes _hover01" id="exitYes">
-                <a href="modal/signin.html">Да, выйти</a>{" "}
+                <Link to="/login">Да, выйти</Link>{" "}
               </button>
               <button className="pop-exit__exit-no _hover03" id="exitNo">
-                <a href="main.html">Нет, остаться</a>{" "}
+                <Link to="/">Нет, остаться</Link>{" "}
               </button>
             </div>
           </form>
