+++ vite-project/src/pages/Logout.jsx
@@ -1,4 +1,6 @@
-export function MainPage () {
+import { PopUser } from "../components/popUps/popUser/popUser.jsx"
+
+export function Logout () {
     return (
         <div>
             Logout page
